optimizacion del codigo, pruebas realizadas con postman. Agregados me…

+++ api/src/controllers/Post.js
@@ -57,15 +57,23 @@ const getPost = (req, res, next) => {
 };
 
 const addPost = async (req, res, next) => {
+  const { idUser } = req.params;
+  const { title, message, tag, module } = req.body;
   try {
-    const { idUser } = req.params;
     const createdBy = await User.findByPk(idUser);
+    
+    const exist = await Post.findAll({
+      where: {
+        title,
+        message
+      }
+    });
 
-    const { title, message, rating, tag, module } = req.body;
+    if(exist.length) return res.status(400).send("Rechazado, ya existe un posteo con el mismo titulo y mensaje")
+
     const postCreated = await Post.create({
       title,
-      message,
-      rating,
+      message
     });
     const tags = await Tag.findAll({
       where: {
@@ -80,26 +88,39 @@ const addPost = async (req, res, next) => {
     postCreated.addTag(tags);
     section[0].addPost(postCreated);
     createdBy.addPost(postCreated);
-    return res.send("Post done successfully");
+    res.json(postCreated);
   } catch (error) {
     next(error);
   }
 };
 
 const updatePost = async (req, res, next) => {
-  const { idPost } = req.params;
-  const { title, message, rating, tag } = req.body;
+  const { idPost, idUser } = req.params;
+  const { title, message, tag } = req.body;
   try {
-    const postUpdate = await Post.findByPk(idPost);
-    const allTags = await Tag.findAll({
-      where: {
-        name: tag,
-      },
-    });
-    console.log(allTags);
-    await postUpdate.setTags(allTags);
+    const postUpdate = await Post.findByPk(idPost, { include: [User, Module] });
+    const ownUser = await User.findByPk(idUser)
+
+    if(!postUpdate || !ownUser) return res.status(404).send("Datos no encontrados")
+
+    if(ownUser.id !== postUpdate.dataValues.user.id) {
+      return res.status(400).send("Accion denegada, solo el propietario puede actualizar el post")
+    }
+
+    if(tag) {
+      const allTags = await Tag.findAll({
+        where: {
+          name: tag,
+        },
+        include: [Module]
+      });
+      let pass = allTags.find(e => e.dataValues.module.name !== postUpdate.dataValues.module.name)
+      if(Boolean(pass)) return res.status(400).send(`Tags deben pertenecer a Modulo ${postUpdate.dataValues.module.name}`)
+      await postUpdate.setTags(allTags);
+    }
+
     const updateSuccess = await Post.update(
-      { title, message, rating },
+      { title, message },
       {
         where: { id: idPost },
         raw: true,
@@ -113,15 +134,26 @@ const updatePost = async (req, res, next) => {
   }
 };
 
-const deletePost = (req, res, next) => {
-  const { idPost } = req.params;
-  return Post.destroy({
-    where: {
-      id: idPost,
-    },
-  })
-    .then(() => res.status(200).send("post deleted successfully"))
-    .catch((error) => next(error));
+const deletePost = async(req, res, next) => {
+  const { idPost, idUser } = req.params;
+  try {
+    const post = await Post.findByPk(idPost, { include: [User] });
+    const user = await User.findByPk(idUser)
+
+    if(!post || !user) return res.status(404).send("Datos no encontrados")
+
+    if(post.dataValues.user.id === user.id || user.isAdmin) {
+      await Post.destroy({
+        where: {
+          id: idPost
+        }
+      });
+    } else return res.status(400).send("Accion denegada, solo el propietario puede eliminar el posteo")
+    
+    res.send("Post deleted successfully")
+  } catch (error) {
+    next(error)
+  }
 };
 
 module.exports = {
